CI: simplify with make ci-archive and use kdevops-results-archive
Upstream kdevops now has support for 'make ci-archive' which simplifies
the ways in which each repository tree which wants a CI archive can
define it and use it to build not the zip artifact we want but and also
an xz archive which we upload into kdevops-results-archive [1].

The kdevops-results-archive uses git LFS which let's users easily
track *all* results from kdevops in a light way, ie, without git LFS
users would download currently 210M and with git LFS and using the
GIT_LFS_SKIP_SMUDGE=1 environment variable they only download 1.6M
for a total space savings of 99.24%. With git LFS you only download
file pointers for the large files you want to track, and can optionally
download the files you need when you need them.

This lets us keep a unified archive of all results even if they expire
as artifacts on public CI infrastructure like github and gitlab.

The experience we have so is that the less a github runner does the
better, and so the 'make ci-archive' heavily simplifies the dynamic
way in which a repository defines the files they need to track.

[0] https://github.com/linux-kdevops/kdevops-results-archive

Signed-off-by: Luis Chamberlain <mcgrof@kernel.org>