CI: handle metadata for kernel hacking: ci.subject special handling
We want the metadata from the Linux git tree, not the kdevops tree,
so cd to kdovops/linux, not kdevops.

Additionally there are two ways to leverage github actions for kdevops
for Linux kernel development:

  a) Let kernel-patch-daemon push series for you onto a tree
  b) You push an existing development branch to test yourself

For a), it merges the .github  directory *prior* to putting
the patches which patchwork picks up on, and in that case it
makes sense to use the last commit for the ci.subject.

For b) it makes sense for developers to have prepared kernel
git trees where we don't want to modify the git tree, and so
for that case it makes sense to merge the commits which add the
.github directory as the last commit. So we special-case that,
and look for that.

Signed-off-by: Luis Chamberlain <mcgrof@kernel.org>