CI: use ci.ref
If a kernel maintainer is going use github actions, then the last
commit can add the .github directory and kernel-patch-daemon can be
skipped. In this case we want to use a ref which is relevant towards
development. So detect if the last commit added the .github directory
and if so inform kdevops of this using the new optional ci.ref file
which if present will allow us to use that ref and associated tag
if present for the refer on the kdevops-results-archive commit log
and archive.

Signed-off-by: Luis Chamberlain <mcgrof@kernel.org>