CI: split work files
Striving to share *one* .github/workflows/kdevops.yml file gets sloppy
once we consider possible variability we want to support for the option
input menu. Yes, its possible to support one file to share all that, but
that involves spinning a docker container just to check the github repo
name and enable or disable certain files. Instead what we can do is
share *as much as we can* into the initialization and cleanup and then
we have topic branches inside this tree for each major subsystems. The
main branch can track advances for the files:

  - kdevops-init.yml     - Used for initialization
  - kdevops-cleanup.yml  - Used at cleanup time
  - kdevops-generic.yml  - Example of how to use these

The kdevops-generic.yml can also cover most simple subsystems. And so
then topic branches can be created based on the main branch.

The development for these 3 files can go on in the main branch and
topic branches can be updated with time.

Signed-off-by: Luis Chamberlain <mcgrof@kernel.org>