kdevops-init.yml: add a relevant commit and use it all over
There are two uses case for CI testing with kdevops:

1) kernel-patch-daemon  [0] which will merge the kdevops-ci topic
   branch onto a kernel tree as the *first* commit and then merge a
   series of patches from patchwork and then create a pull request
   onto a repository so to trigger .github actions.

2) A kernel developer just merges the relevant topic branch from
   kdevops-ci onto their kernel git tree as their *last* commit
   so to maintain existing git tree sha1sums and tags.

For the use case 2) we want to do a bit of homework on the relevant
commit. We just parse if the last commit added a .github directory
and if so assume that's not a relevant commit and get the one prior.

Either way, we collect a tag in case we can find one as those are
pleasant.

We can use *this* ref to the actual build process test and also the
kdevops configuration, this doesn't change much other than making
things clearer for the second use case, and it also makes the relevant
ref tag be used for kdevops guests being built.

Link: https://github.com/facebookincubator/kernel-patches-daemon # [0]
Signed-off-by: Luis Chamberlain <mcgrof@kernel.org>