kdevops-fstests.yml: work around github dynamic timeout variable bug
If we try to use a dynamic timeout it works but if you try to propagate
it as an output variable an error is detected even though the error
message clearly sees the variable we set. If we simply replace the
variable with a hard coded value though this is fixed. So even though
we know we are limited to only 5 days, let's hard code a value of 1
month max time, knowing already that we did input processing for
workflow dispatch to only allow max 5 days.

If the tasks go over 5 days, github will terminate jobs for us.

Signed-off-by: Luis Chamberlain <mcgrof@kernel.org>