kdevops-cleanup.yml: respect manual cancellation
If the user cancels a run let's respect that otherwise we end
up spinning waiting until the last valid action completes just
to try to cancel the run, but it'll run the make ci-archive and
make destroy. We can safely skip 'make ci-archive' but we do
want to try to run 'make destroy' regardless so we clean up the
possible guests which may have been instantiated.

If we really *want* this later, we just need to document that
users should only hit cancel if they are ready to log into systems
and clear the guests themselves. But since that may not be the
case for all self-hosted servers, waiting to clear up the guests
is needed for now.

Signed-off-by: Luis Chamberlain <mcgrof@kernel.org>