kdevops-init.yml: add initial ssh-agent sanity check
Since we are using the webfactory ssh-agent to connect to
the kdevops-results-archive to push results we want to fail
right away instead of the last minute so we can debug ssh
key issues early on, not after a long test.

So add an early sanity test for connectivity.

Also ensure we add StrictHostKeyChecking otherwise we see the
run.sh is waiting for us to hit "yes" on the prompt.

Signed-off-by: Luis Chamberlain <mcgrof@kernel.org>