fstests: process selected tests and soak
Although the menu had gotten support to click and select soak duration
and the custom set of tests we could use we never propagated this down
to the init for one reason: trying to figure out if we can share this
with the other CI runners.

I don't see how to do this cleanly so just diverge a bit. We can always
accept innovations on the kdevops-init.yml and kdevops-cleanup.yml files
into the fstests branch with just:

  git merge main

And then manually editing the delta, and in practice this just requires
removing the kdevops-generic.yml file, and manually adjusting the changes
we actually care about. Accept this as life and move on.

Signed-off-by: Luis Chamberlain <mcgrof@kernel.org>