CI: leverage fstests-baseline-loop
The kdevops-baseline-loop was originally developed prior to kdevops's
support for fstest soak duration, and we used to rely on a test steady
state of 100, so we'd run fstsets in a loop 100 times.

With soak duration supported, we can just worry about letting the
developers pick the right soak duration when needed, and since the
steady state goal is now just 1 as the default, and since kdevops
CI has always been enabled to just expose the target:

  * make fstests-baseline-loop

In effect this target now is just enabling the watchdog, and
the existing CI is doing something similar. Since the watchdog
is already being used by fstests-baseline-loop and we output
the status on .kernel.ci.status we can just cat that file now
every 60 seconds too.

We also collect the .kernel.ci.* files.

This will let us test if we can pick up a crash now.

Signed-off-by: Luis Chamberlain <mcgrof@kernel.org>