kdevops-fstests.yml: collect results again in case of a crash
kdevops now enables FSTESTS_WATCHDOG_RESET_HUNG_SYSTEMS by default
and when 'make fstests-baseline-loop' is run hung systems may be reset
but in order for the CI to finish the ansible tasks may be terminated
as well, and so the last step of result collection won't run, so we
don't "pas go and collect $200". To fix this, ensure we always just
run 'make fstests-results' as the last step. We don't loose much
in running it again if it already worked once. Later on we can perhaps
detect if this is needed but since we want to move on with life just
shoe-horn this in.

Signed-off-by: Luis Chamberlain <mcgrof@kernel.org>