CI: support running the same test on the same system
Early when working with a CI system with kdevops we ran into issues
with several tests running into conflicts with the same prefix on the
same system. We fixed this by using the 12 character git commit ID of
the target kernel we're testing.

We figured we would not run into the situation where we would want to
run the same test against the same subsystem for at the same time.
Support for kpd with a lei setup seems to show that this is a requirement
and we didn't need to support this before.

Konstantin help set up a patchwork instance that uses lei to customize
which patches get queued up, instead of a full mailing list being
monitored only select patches get considered. This helps speedup
testing as we only focus on things we care about for busy work on CIs.

The first one we have is for the firmware loader subsystem which I help
maintain:

https://patchwork.kernel.org/project/firmware/list/

The respective github actions are:

https://github.com/linux-kdevops/linux-firmware-kpd/actions/

I improved docs for kdevops CI here:

https://github.com/linux-kdevops/kdevops/blob/main/docs/kernel-ci/README.md

The KPD part here is here:

https://github.com/linux-kdevops/kdevops/blob/main/docs/kernel-ci/kernel-ci-kpd.md

I see patches keep piling up for the firmware-loader lei based patchwork
but my kpd instance has run into trouble with actions failing.

I've debugged the issues, and narrowed the issues down to first a stupid
systemd bug with systemd-networkd-wait-online.service which I now
resolved after 2 attempts:

https://github.com/linux-kdevops/kdevops/commit/683610d65556377f213df980d5256bee4d60c467
https://github.com/linux-kdevops/kdevops/commit/b9e7570e72dac2df62d2ba80b01e3ff534316517

Now the last remaining bug is caused because of a limitation of kdevops
to not be able to run guests with the same name at the same time.
This was rather a design choice because obviously you cannot have two
guests of the same type run at the same time, and I wanted for the guest
names that come up to reflect exactly the commit ID being tested along
with the test name. The assumption was that the same host would not run
the same type of test twice on the same system for same commit ID.

kpd bringup with a lei seems to be a corner case.

On bring up it creates 3 separate branches to ramp up testing in parallel:

test  order |  kpd commit name                  | tree
1           |  adding ci files                  | driver-core-next_base
2           |  Dummy comit                      | driver-core-next_test
3           | [test] driver-core-next_test #38  | driver-core-next_test

This fails because they are all trying to create the guest with name:

acfeb6defcb9-firmware

We could add something unique like the ${{ github.run_id }} but we need
to separate the commit ID from the ${{ github.run_id }}. I first tried
a simple "." but if while hostname may return the right hostname, $HOST
will only be set to the first part, and so some scripts which rely on
$HOST may not work.

I tried other odd things like UTF characters but those don't work at all
in many cases. To resolve this in a way that just works with the
existing scripts just add KDEVOPS to separate these two.

Signed-off-by: Luis Chamberlain <mcgrof@kernel.org>