kdevops-mm.yml: do not try to run make sefltests-results upon crash
If we detect a crash we bail our tests and then later try to
call make selftests-results but that proves to also have issues.
Likely for example it can happen if the kernel is just failing to
boot properly in the worst case. Let's simplify the collection
process by just having the crash logs.

Signed-off-by: Luis Chamberlain <mcgrof@kernel.org>