.github/actions: add kdevops actions
GitHub Actions allows you to cleanup the workflow directory and share
logic at the job step level using composite actions. However, it
comes with a limitation as the use of 'uses:' inside the action is not
allowed. Any step that rely on uses need to go out into the workflow
file.

Signed-off-by: Daniel Gomez <da.gomez@samsung.com>